backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi[standard]>=0.114.2,<1.0.0",
    "python-multipart<1.0.0,>=0.0.7",
    "email-validator<3.0.0.0,>=2.1.0.post1",
    "tenacity<9.0.0,>=8.2.3",
    "pydantic>2.0",
    "alembic<2.0.0,>=1.12.1",
    "httpx<1.0.0,>=0.25.1",
    "psycopg[binary]<4.0.0,>=3.1.13",
    "psycopg-pool>=3.2.0",
    "sqlmodel>=0.0.24",
    "pydantic-settings<3.0.0,>=2.2.1",
    "pyjwt<3.0.0,>=2.8.0",
    "bcrypt<5.0.0,>=4.1.0",
    "PyYAML==6.0.2",
    "kubernetes>=33.1.0",
    "cachetools>=5.0.0",
    "boto3>=1.35.0",
    "google-cloud-storage>=2.0.0",
    "langchain-openai>=0.2.0",
]

[tool.uv]
dev-dependencies = [
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"